% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntaxLchange.R
\name{model_syntax_change}
\alias{model_syntax_change}
\title{Generate model syntax for a latent change score model}
\usage{
model_syntax_change(
  waves = 10,
  model_type = c("latent_change_x", "latent_change_xy")
)
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}

\item{model_type}{Specify whether estimating a latent change score model for one variable or two.}
}
\value{
character string containing the Lavaan model syntax.
}
\description{
Generate model syntax for a latent change score model
}
