% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntaxCLPM.R
\name{model_syntax_clpm}
\alias{model_syntax_clpm}
\title{Generate model syntax for a cross-lagged panel model, with or without random intercepts}
\usage{
model_syntax_clpm(waves = 10, model_type = c("clpm", "ri-clpm"))
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}

\item{model_type}{Specify whether an RI-CLPM or CLPM.}
}
\value{
character string containing the Lavaan model syntax.
}
\description{
Generate model syntax for a cross-lagged panel model, with or without random intercepts
}
