% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRICLPM.R
\name{simulate_riclpm}
\alias{simulate_riclpm}
\title{simulate_riclpm}
\usage{
simulate_riclpm(
  waves = 10,
  stability.p = 0.2,
  stability.q = 0.2,
  cross.p = 0.1,
  cross.q = 0.1,
  variance.p = 3,
  variance.q = 3,
  cov.pq = 0.1,
  variance.between.x = 1,
  variance.between.y = 1,
  cov.between = 0.5,
  ...
)
}
\arguments{
\item{waves}{The number of waves (time points) in the model.}

\item{stability.p}{The stability parameter for the x variable (autoregressive effect).}

\item{stability.q}{The stability parameter for the y variable (autoregressive effect).}

\item{cross.p}{The cross-lagged effect of y on x at the next time point.}

\item{cross.q}{The cross-lagged effect of x on y at the next time point.}

\item{cov.pq}{The covariance between x and y within the same time point.}

\item{...}{Additional arguments to pass to the `lavaan::simulateData` function.}
}
\value{
A list containing two elements:
   * `model`: The Lavaan model syntax used for data simulation.
   * `data`:  The simulated data in a data frame format.
}
\description{
Simulate data from an observed, random intercept cross-lagged regression (RI-CLR) model
In particular, this creates a synthetic dataset of a cross-lagged model with a specified number of waves and structural parameters.
}
