% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.data1.R
\docType{data}
\name{test_data1}
\alias{test_data1}
\title{Data for the first time point used for testing}
\format{
A dataframe.
}
\source{
Generated from case information of a hospital in Hebei province
}
\description{
Data for the first time point used for testing
}
\examples{
data(test_data1)
}
