% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.data2.R
\docType{data}
\name{test_data2}
\alias{test_data2}
\title{Data for the second time point used for testing}
\format{
A dataframe.
}
\source{
Generated from case information of a hospital in Hebei province
}
\description{
Data for the second time point used for testing
}
\examples{
data(test_data2)
}
